requirements.txt
django==4.0.1
djangorestframework==3.13.1
drf-yasg==1.20.0
djangorestframework-simplejwt==5.0.0
django_cors_headers==3.11.0
django-filter==21.1